+++ programs/x509/load_roots.c
@@ -82,7 +82,7 @@ int main(int argc, char *argv[])
     int exit_code = MBEDTLS_EXIT_FAILURE;
     unsigned i, j;
     struct mbedtls_timing_hr_time timer;
-    unsigned long ms;
+    unsigned long long ms;
 
     psa_status_t status = psa_crypto_init();
     if (status != PSA_SUCCESS) {
@@ -151,7 +151,7 @@ int main(int argc, char *argv[])
         mbedtls_printf(".");
     }
     ms = mbedtls_timing_get_timer(&timer, 0);
-    mbedtls_printf("\n%u iterations -> %lu ms\n", opt.iterations, ms);
+    mbedtls_printf("\n%u iterations -> %llu ms\n", opt.iterations, ms);
     exit_code = MBEDTLS_EXIT_SUCCESS;
 
 exit:
